+++ public/index.html
@@ -33,8 +33,8 @@
 
     <button class="clear">Clear</button>
 
-<script src="bundle.js"></script>
 
+    <script src="bundle.js"></script>
     <!-- <script src="/js/add.js"> </script>
     <script src="/js/todo.js"> </script> -->
 </body>
